Added mapper to map fetched game to our game SteamImage and Game ID changed from int to long rename variables and add types

+++ PriceTrackerMobile/Helpers/SteamImage.cs
@@ -2,7 +2,7 @@
 {
     class SteamImage
     {
-        public static string FromId(int id)
+        public static string FromId(long id)
         {
             return $"https://cdn.akamai.steamstatic.com/steam/apps/{id}/header.jpg";
         }
